ml/requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
beautifulsoup4==4.14.3
bleach==6.2.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
gdown==5.2.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.13
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kaggle==1.7.4.5
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnx==1.19.1
onnxruntime==1.19.2
opencv-python==4.13.0.92
packaging==26.1
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.26
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.3
typer==0.23.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
webencodings==0.5.1
zipp==3.23.1
